tests/wpt/tests/svg/text/parsing/shape-inside-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Text: parsing shape-inside with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/text.html#ShapeInsideProperty"/>
    <h:meta name="assert" content="shape-inside supports the full grammar 'auto | [ basic-shape | uri ]+'"/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("shape-inside", "auto");

test_valid_value("shape-inside", "circle(10px at 20px -30px)");
test_valid_value("shape-inside", "ellipse(10px 20px at -30px 40px)");
test_valid_value("shape-inside", "polygon(10px 20px, -30px 40px, 50px 60px)");

test_valid_value("shape-inside", 'url("https://example.com/")');

test_valid_value("shape-inside", 'circle(10px at 20px -30px) url("https://example.com/") ellipse(10px 20px at -30px 40px)');

  ]]></script>
</svg>
